python annotations

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -4,9 +4,9 @@ import typing
 from typing import List
 
 
-def sum_mixed_list(mxd_list: List[typing.Union[float, int]]) -> float:
+def sum_mixed_list(mxd_list: List[typing.Union[int, float]]) -> float:
     """sums the mixed list"""
     sumofmixedlist: float = 0
     for num in mxd_list:
         sumofmixedlist += num
-    return sumofmixedlist
+    return float(sumofmixedlist)
